EventStudies

Documentation for EventStudies.

EventStudies.AbstractEventStudyModel

EventStudies.AugmentedMarketModel

EventStudies.BootstrapInference

EventStudies.ClassicInference

EventStudies.ConstantMeanReturn

EventStudies.DataMissing

EventStudies.EventStatus

EventStudies.ExcessReturn

EventStudies.InvalidColname

EventStudies.MarketModel

EventStudies.ModelDataMissing

EventStudies.ModelIndexIncompatible

EventStudies.NoModel

EventStudies.Success

EventStudies.WilcoxonInference

EventStudies.WrongSpan

EventStudies.apply_model

EventStudies.inference

EventStudies.inference

EventStudies.inference

EventStudies.inference_surface

EventStudies.levels_to_returns

EventStudies.remap_cumsum

EventStudies.remap_levels

EventStudies.to_eventtime_windowed

EventStudies.AbstractEventStudyModel — Type

EventStudies.AugmentedMarketModel — Type

AugmentedMarketModel(market_returns::TSFrame, additional_returns::TSFrame)

Applies the augmented market model of eventstudy.R to the data.

EventStudies.BootstrapInference — Type

BootstrapInference(sampling_method::Bootstrap.BootstrapSampling)
BootstrapInference(; replicates = 1000)

Creates an instance of BootstrapSampling for use in inference. This can be constructed using a Bootstrap.BootstrapSampling subtype directly, or by the convenience keyword constructor which creates a BootstrapInference backed by BasicSampling.

Available sampling methods

Random resampling with replacement (BasicSampling)

Antithetic resampling, introducing negative correlation between samples (AntitheticSampling)

Balanced random resampling, reducing bias (BalancedSampling)

Exact resampling, iterating through all unique resamples ([]ExactSampling](@ref)): deterministic bootstrap, suited for small samples sizes

Resampling of residuals in generalized linear models (ResidualSampling, WildSampling)

Maximum Entropy bootstrapping for dependent and non-stationary datasets (MaximumEntropySampling)

EventStudies.ClassicInference — Type

ClassicInference()

Creates an instance of ClassicInference for use in inference. This forwards to the T-test from HypothesisTests.jl.

EventStudies.ConstantMeanReturn — Type

ConstantMeanReturn()

Each column is decreased by its mean.

EventStudies.DataMissing — Type

This means that there was missing data within the window.

EventStudies.EventStatus — Type

abstract type EventStatus

An abstract supertype for all event status types. These are used to indicate the status of an event (i.e., if it was successful or not, and if so why it was unsuccessful).

EventStudies.ExcessReturn — Type

ExcessReturn(market_returns::TSFrame)

This models basic scalar excess return.

EventStudies.InvalidColname — Type

This means that the column name was not found in the data. It prints the column name. Equivalent to unitmissing in eventstudies.R.

EventStudies.MarketModel — Type

MarketModel(market_returns::TSFrame)

Regresses "market" returns on the variable using a linear model.

EventStudies.ModelDataMissing — Type

This means that the model was missing data within the event time window.

EventStudies.ModelIndexIncompatible — Type

This means that the model's index type was incompatible with the event's index type.

EventStudies.NoModel — Type

NoModel()

A very fancy way to say that no decorrelation needs to be done. No-op.

EventStudies.Success — Type

Represents a successful event

EventStudies.WilcoxonInference — Type

WilcoxonInference(; exact = nothing)

Creates an instance of WilcoxonInference for use in inference. This forwards to the Wilcoxon inference test from HypothesisTests.jl.

If exact = nothing, then a heuristic determines whether to use the exact or approximate Wilcoxon signed-rank test. If exact = true, then the exact test is used; if exact = false, then the approximate test is used. If exact = nothing, then a heuristic determines whether to use the exact or approximate tests.

EventStudies.WrongSpan — Type

This means that the span of the window extended outside the available data.

EventStudies.apply_model — Method

apply_model(model::AbstractEventStudyModel, data::TSFrame)::(::TSFrame, ::EventStatus)

Applies the given model to the provided TSFrame, and returns a TSFrame (may be empty) along with an EventStatus code. If the EventStatus code is not Success(), then the returned TSFrame should (and will) be ignored.

EventStudies.inference — Function

inference(::ClassicInference, ts::TSFrame, conf = 0.975)

Performs classic T-test inference and returns a tuple of vectors (t₀, lower, upper).

EventStudies.inference — Function

inference(inf::BootstrapInference, ts::TSFrame, conf = 0.975)

Performs bootstrap inference and returns a tuple of vectors (t₀, lower, upper).

EventStudies.inference — Method

inference(::WilcoxonInference{ExactType}, ts::TSFrame, conf = 0.975) where ExactType

Performs Wilcoxon signed-rank inference and returns a tuple of vectors (t₀, lower, upper).

EventStudies.inference_surface — Function

inference_surface(infer::InferenceMethod, ts::TSFrame, conf = 0.975; density = 200)::(y_values::Matrix, y_colors::Matrix)

EventStudies.levels_to_returns — Method

levels_to_returns(ts::TSFrame)::TSFrame

Converts the data in ts into "returns" data, i.e., executes diff(log(ts)) .* 100 on each column of ts.

EventStudies.remap_cumsum — Method

remap_cumsum(ts::TSFrame; base = 0)

This function remaps a time series into its cumulative summation. The first value of each column is set to base.

EventStudies.remap_levels — Function

remap_levels(ts::TSFrame, base = 100)

EventStudies.to_eventtime_windowed — Method

to_eventtime_windowed(
    return_timeseries::TSFrame, 
    event_times::Vector{Pair{Symbol, T}}, 
    window::Int, 
    model::AbstractEventStudyModel = NoModel(); 
    debug = false
    ) where T

Takes in a TSFrame of returns, a vector of pairs of column names and event times, and a window size. It then returns a TSFrame of event windows, where each column is the event return for the corresponding event.

Arguments

return_timeseries is a TSFrame which contains the returns of the assets.

event_times is a vector of pairs, where the first element is the name of the column in return_timeseries and the second element is the time of the event. This is in the form :colname => date.

window is the number of data points before and after the event to include in the event window. It may be an Integer, which describes a range -window:1:window. It may also be an AbstractVector of Integers or Dates.Periods, which describes the exact window to use.

model is an AbstractEventStudyModel which is used to transform the data before calculating the event returns.

debug is a boolean which indicates whether to print debug information.