Event time

Event time is the fundamental concept of the event study. When an event study is performed, windows around each event are extracted, and re-indexed to an "event time" where the event occurs at $t=0$.

EventStudies.jl offers a simple interface to perform direct conversion of a timeseries and a set of events to event time, with the physical_to_event_time function:

Missing docstring.

Missing docstring for EventStudies.physical_to_event_time. Check Documenter's build log for details.

Event codes

Each event has an associated event time. If the time is not viable to study (outside the index of the input data), or there is some issue with either the data or the model, then the event will be ignored.

Ignored events have an event code which describes why they were ignored; successful events have the code EventStudies.Success.

EventStudies.DataMissing — Type

This means that there was missing data within the window.

EventStudies.EventStatus — Type

abstract type EventStatus

An abstract supertype for all event status types. These are used to indicate the status of an event (i.e., if it was successful or not, and if so why it was unsuccessful).

EventStudies.InvalidColname — Type

This means that the column name was not found in the data. It prints the column name. Equivalent to unitmissing in eventstudies.R.

EventStudies.MissingAfterModel — Type

This means that the data had a missing value after the model's prediction.

EventStudies.ModelDataMissing — Type

This means that the model was missing data within the event time window.

EventStudies.ModelIndexIncompatible — Type

This means that the model's index type was incompatible with the event's index type.

EventStudies.Success — Type

Represents a successful event

EventStudies.WrongSpan — Type

This means that the span of the window extended outside the available data.