Models

Types of models

EventStudies.Models.AbstractModel — Type

abstract type Models.AbstractModel

Abstract type for models used in EventStudies.jl. Any model should implement: StatsBase.fit!(model, data::TSFrame) and Models.apply(model, data::TSFrame).

EventStudies.Models.ConstantMeanReturn — Type

ConstantMeanReturn()

Each column is decreased by its mean.

EventStudies.Models.ExcessReturn — Type

ExcessReturn(market_returns::TSFrame)

This models basic scalar excess return.

EventStudies.Models.MarketModel — Type

struct MarketModel <: AbstractModel
MarketModel(market_returns::TSFrame)

A model which fits the firm's returns to the market returns, using some provided index or market indicator.

Fit using fit or fit!, and apply to data using apply

Fields

market_returns: A TSFrame which holds the market indicator/index data.

coefs: The coefficients of the fit. Set by fit!.

Model API

To define a new model, you need to create a struct MyModel <: EventStudies.Models.AbstractModel, whose contents can be arbitrary.

Then, you have to implement StatsBase.fit!(model::MyModel, data::TSFrame)::MyModel, which should mutate the model to store the coefficients (you can do this in an immutable struct by making one of the fields a mutable type like a Vector).

Finally, you have to implement StatsBase.predict(model::MyModel, data::TSFrame)::TSFrame which runs the model and returns the result as a TSFrame. Column names should not be changed and data should not be mutated.

Optionally, you can also implement Models.check_window(model::MyModel, window)::Bool, which checks whether the given window exists in your model.

Missing docstring.

Missing docstring for EventStudies.StatsBase.fit!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for EventStudies.StatsBase.predict. Check Documenter's build log for details.